refactor(search-api): DatasetsSearchController - separate out query util class
