+++ applications/search-api/src/main/java/no/fdk/searchapi/controller/datasetssearch/Common.java
@@ -1,0 +1,5 @@
+package no.fdk.searchapi.controller.datasetssearch;
+
+public class Common {
+    public static final String MISSING = "Ukjent";
+}
